feat!: make encoders and validators classes
Breaking Change: This refactors the encoders/validators to be classes which are instantiated and
used.

The Base62Encoder is replaced with a generic `BaseXEncoder` which allows consumers to provider
their own alphabets, if needed.

This change means the exports of the library have changed and so this is a breaking change for
those that relied on some of those exports.