feat: allow consumers to pass in their own PRNG
When constructing a TokenGenerator, a `prng` option can be provided to allow
overriding of the native (`crypto.randomBytes()`) call. This must be a function
that accepts a numeric argument (length of `Buffer` to return) and returns a `Promise`
that resolves to a `Buffer` with the requisite number of random bytes. This is primarily
useful for testing.